fix: move decompose-prompt to DecomposePromptBuilder in src/prompts/builders (M-3)
